SSH: Split internal client to separate file
