+++ plugins/ssh/rust/src/lib.rs
@@ -1,5 +1,6 @@
 use godot::prelude::*;
 
+mod internal_ssh_client;
 mod ssh_client;
 
 struct DreamDeckSSH;
